refactor(keystone-operator): split reconcile_database.go into four files
reconcile_database.go had reached 901 lines spanning four distinct
concerns. Split it along those seams, moving each block verbatim:

  - db_job_metrics.go    — the db_sync metric dedupe (dbJobUID* +
                           recordDBJobTerminalState, CC-0089).
  - reconcile_upgrade.go — the expand-migrate-contract upgrade state
                           machine (isUpgrade, initiateUpgrade, the phase
                           reconcilers and their condition helpers, CC-0056).
  - database_jobs.go     — the MariaDB CR and db_sync Job builders.

reconcile_database.go keeps the core managed/brownfield reconcile, the
MariaDB CR identity (mariaDBResourceCtors / mariaDBResourceKey), the
DatabaseReady reason constants, the cluster-readiness probe, and the
finalizer. All symbols keep their names and the moves are within the same
package, so every existing unit test exercises them unchanged.

Pure refactor; no behavior change.

AI-assisted: Claude Code
On-behalf-of: @SAP christian.berendt@sap.com
