+++ operators/keystone/internal/controller/reconcile_database.go
@@ -13,7 +13,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
@@ -22,27 +21,7 @@ import (
 	"github.com/c5c3/forge/internal/common/database"
 	"github.com/c5c3/forge/internal/common/job"
 	keystonev1alpha1 "github.com/c5c3/forge/operators/keystone/api/v1alpha1"
-	"github.com/c5c3/forge/operators/keystone/internal/metrics"
 )
-
-// dbJobUIDAnnotationFormat builds the annotation key used to dedupe terminal
-// metric emission for a given DB-related Job. The annotation lives on the
-// Keystone CR so it survives Job deletion (e.g. when RunJob recreates the Job
-// after a PodSpec hash change) and the operator can still skip re-emission
-// when the same UID is observed twice.
-//
-// The %s placeholder is the buildDBJob nameSuffix ("db-sync", "db-expand",
-// "db-migrate", "db-contract", "schema-check"). Each phase keeps an
-// independent dedupe annotation so metric emission for the expand-migrate-
-// contract upgrade Jobs and the post-sync schema-check Job is never silently
-// suppressed by an unrelated phase having already been observed.
-const dbJobUIDAnnotationFormat = "forge.c5c3.io/last-%s-job-uid"
-
-// dbJobUIDAnnotationKey returns the dedupe annotation key for the Job
-// identified by the given nameSuffix (see dbJobUIDAnnotationFormat).
-func dbJobUIDAnnotationKey(jobSuffix string) string {
-	return fmt.Sprintf(dbJobUIDAnnotationFormat, jobSuffix)
-}
 
 // mariaDBResourceCtors is the single source of truth for the set of MariaDB CR
 // kinds owned by a Keystone CR. Both the finalizer cleanup
@@ -101,132 +80,6 @@ const (
 // these (and not db-sync or schema-check, which belong to the steady-state path)
 // so reverting an in-flight upgrade leaves no stale phase Jobs behind (#468).
 var upgradeJobSuffixes = []string{"db-expand", "db-migrate", "db-contract"}
-
-// recordDBJobTerminalState observes the named DB-related Job's terminal
-// condition and emits keystone_operator_db_sync_total /
-// keystone_operator_db_sync_duration_seconds exactly once per (Job suffix, Job
-// UID) tuple. The jobSuffix argument is the buildDBJob
-// nameSuffix that identifies the phase ("db-sync", "db-expand", "db-migrate",
-// "db-contract", "schema-check"); the same suffix selects the per-phase
-// dedupe annotation via dbJobUIDAnnotationKey, so terminal emission for one
-// phase never silently suppresses another. The function is best-effort: any
-// error reading the Job or persisting the UID annotation is swallowed rather
-// than surfaced as a reconcile failure, because the metric is an observability
-// signal that must not block the sub-reconciler.
-//
-// Idempotence: the last-observed Job UID is stored as a Keystone CR
-// annotation. Re-observing the same UID skips emission; a Job recreated by
-// RunJob (e.g. after a PodSpec hash change) carries a fresh UID and therefore
-// drives a fresh metric observation. Duration is computed as
-// condition.LastTransitionTime minus Job.CreationTimestamp.
-//
-// Ordering the dedupe annotation is patched
-// BEFORE metrics.RecordDBSync. If the patch fails (transient apiserver error,
-// stale-RV conflict), the metric is NOT emitted on this pass — the next
-// reconcile re-evaluates the Job and either emits then (after a successful
-// patch) or surfaces a single deferred-emission log line. This preserves the
-// at-most-once-per-UID guarantee documented in
-// docs/reference/keystone-operator-metrics.md against transient apiserver
-// failures, at the cost of postponing the observation by one reconcile.
-//
-// Visibility on persistent patch failure (review-2 suggestion 1):
-// the deferred-emission log is emitted at Info (default level) and a
-// Warning event is recorded on the Keystone CR. A persistent apiserver
-// Patch failure would otherwise silently degrade db_sync metric emission
-// to at-most-once-per-UID-per-successful-patch — invisible at default log
-// levels until an alert fires for missing samples. Emitting at Info plus a
-// CR-visible event ensures cluster operators notice the degradation
-// directly via `kubectl describe keystone` and default-verbosity logs.
-func (r *KeystoneReconciler) recordDBJobTerminalState(ctx context.Context, keystone *keystonev1alpha1.Keystone, jobSuffix string) {
-	annotationKey := dbJobUIDAnnotationKey(jobSuffix)
-	jobName := fmt.Sprintf("%s-%s", keystone.Name, jobSuffix)
-	var dbJob batchv1.Job
-	if err := r.Get(ctx, types.NamespacedName{
-		Name:      jobName,
-		Namespace: keystone.Namespace,
-	}, &dbJob); err != nil {
-		return
-	}
-
-	var (
-		result       string
-		transitionAt metav1.Time
-	)
-	for _, c := range dbJob.Status.Conditions {
-		if c.Status != corev1.ConditionTrue {
-			continue
-		}
-		switch c.Type {
-		case batchv1.JobComplete:
-			result = "succeeded"
-			transitionAt = c.LastTransitionTime
-		case batchv1.JobFailed:
-			result = "failed"
-			transitionAt = c.LastTransitionTime
-		case batchv1.JobSuspended, batchv1.JobFailureTarget, batchv1.JobSuccessCriteriaMet:
-			// Non-terminal transitions: a Suspended Job may be resumed and
-			// FailureTarget/SuccessCriteriaMet flip briefly before the
-			// terminal JobFailed/JobComplete is set. Ignore here — the
-			// metric is emitted only on terminal transitions.
-		}
-		if result != "" {
-			break
-		}
-	}
-	if result == "" {
-		return
-	}
-
-	uid := string(dbJob.UID)
-	if uid != "" && keystone.Annotations[annotationKey] == uid {
-		return
-	}
-
-	// Persist the observed UID BEFORE recording the metric so a transient
-	// patch failure cannot cause double-counting on the next reconcile
-	// Patching the caller's keystone directly
-	// would let the client overwrite its in-memory status conditions
-	// (accumulated by earlier sub-reconcilers on this pass) with the
-	// backend copy, so the patch operates on a separate DeepCopy. The
-	// annotation and post-patch ResourceVersion are mirrored back onto the
-	// caller's object so a later Status().Update does not hit a stale-RV
-	// conflict.
-	patchTarget := keystone.DeepCopy()
-	patch := client.MergeFrom(patchTarget.DeepCopy())
-	if patchTarget.Annotations == nil {
-		patchTarget.Annotations = make(map[string]string)
-	}
-	patchTarget.Annotations[annotationKey] = uid
-	if err := r.Patch(ctx, patchTarget, patch); err != nil {
-		// review-2 suggestion 1: log at Info (default level) and
-		// emit a Warning event on the Keystone CR so persistent apiserver
-		// Patch failures are visible without raising log verbosity. A
-		// silent V(1) line could let the at-most-once-per-UID degradation
-		// persist until alerts fire for missing samples.
-		log.FromContext(ctx).Info(
-			"recordDBJobTerminalState: patching last-observed Job UID failed; "+
-				"db_sync metric emission deferred to the next reconcile",
-			"keystone", keystone.Name,
-			"jobSuffix", jobSuffix,
-			"err", err.Error(),
-		)
-		r.Recorder.Eventf(keystone, corev1.EventTypeWarning, "DBSyncMetricEmissionDeferred",
-			"Patching last-observed %s Job UID failed; db_sync metric emission deferred to the next reconcile: %v",
-			jobSuffix, err)
-		return
-	}
-	if keystone.Annotations == nil {
-		keystone.Annotations = make(map[string]string)
-	}
-	keystone.Annotations[annotationKey] = uid
-	keystone.ResourceVersion = patchTarget.ResourceVersion
-
-	duration := transitionAt.Sub(dbJob.CreationTimestamp.Time)
-	if duration < 0 {
-		duration = 0
-	}
-	metrics.RecordDBSync(keystone.Name, keystone.Namespace, result, duration)
-}
 
 // isMariaDBClusterReady returns whether the MariaDB cluster referenced by
 // spec.database.clusterRef currently reports a Ready=True condition. It
@@ -421,100 +274,6 @@ func (r *KeystoneReconciler) reconcileDatabase(ctx context.Context, keystone *ke
 	return ctrl.Result{}, nil
 }
 
-// isUpgrade returns true when spec.image.tag differs from status.installedRelease
-// and the change requires the expand-migrate-contract upgrade flow.
-// Returns false for fresh deployments (empty installedRelease), same version,
-// and patch-only changes.
-func isUpgrade(keystone *keystonev1alpha1.Keystone) bool {
-	if keystone.Status.InstalledRelease == "" {
-		return false
-	}
-	if keystone.Spec.Image.Tag == keystone.Status.InstalledRelease {
-		return false
-	}
-
-	from, err := ParseRelease(keystone.Status.InstalledRelease)
-	if err != nil {
-		// Let initiateUpgrade handle the error with proper conditions.
-		return true
-	}
-	to, err := ParseRelease(keystone.Spec.Image.Tag)
-	if err != nil {
-		return true
-	}
-
-	return !IsPatchOnly(from, to)
-}
-
-// initiateUpgrade validates the upgrade path and sets the initial upgrade state.
-func (r *KeystoneReconciler) initiateUpgrade(ctx context.Context, keystone *keystonev1alpha1.Keystone) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-
-	from, err := ParseRelease(keystone.Status.InstalledRelease)
-	if err != nil {
-		conditions.SetCondition(&keystone.Status.Conditions, metav1.Condition{
-			Type:               "DatabaseReady",
-			Status:             metav1.ConditionFalse,
-			ObservedGeneration: keystone.Generation,
-			Reason:             conditionReasonVersionParseError,
-			Message:            fmt.Sprintf("failed to parse installed release %q: %v", keystone.Status.InstalledRelease, err),
-		})
-		r.Recorder.Eventf(keystone, corev1.EventTypeWarning, conditionReasonVersionParseError, "Failed to parse installed release %q: %v", keystone.Status.InstalledRelease, err)
-		return ctrl.Result{}, fmt.Errorf("parse installed release %q: %w", keystone.Status.InstalledRelease, err)
-	}
-
-	to, err := ParseRelease(keystone.Spec.Image.Tag)
-	if err != nil {
-		conditions.SetCondition(&keystone.Status.Conditions, metav1.Condition{
-			Type:               "DatabaseReady",
-			Status:             metav1.ConditionFalse,
-			ObservedGeneration: keystone.Generation,
-			Reason:             conditionReasonVersionParseError,
-			Message:            fmt.Sprintf("failed to parse target release %q: %v", keystone.Spec.Image.Tag, err),
-		})
-		r.Recorder.Eventf(keystone, corev1.EventTypeWarning, conditionReasonVersionParseError, "Failed to parse target release %q: %v", keystone.Spec.Image.Tag, err)
-		return ctrl.Result{}, fmt.Errorf("parse target release %q: %w", keystone.Spec.Image.Tag, err)
-	}
-
-	if IsDowngrade(from, to) {
-		conditions.SetCondition(&keystone.Status.Conditions, metav1.Condition{
-			Type:               "DatabaseReady",
-			Status:             metav1.ConditionFalse,
-			ObservedGeneration: keystone.Generation,
-			Reason:             conditionReasonDowngradeNotSupported,
-			Message:            fmt.Sprintf("downgrade from %s to %s is not supported", from.Raw, to.Raw),
-		})
-		r.Recorder.Eventf(keystone, corev1.EventTypeWarning, conditionReasonDowngradeNotSupported, "Downgrade from %s to %s is not supported", from.Raw, to.Raw)
-		return ctrl.Result{}, fmt.Errorf("downgrade from %s to %s is not supported", from.Raw, to.Raw)
-	}
-
-	if !IsSequentialUpgrade(from, to) {
-		conditions.SetCondition(&keystone.Status.Conditions, metav1.Condition{
-			Type:               "DatabaseReady",
-			Status:             metav1.ConditionFalse,
-			ObservedGeneration: keystone.Generation,
-			Reason:             conditionReasonUpgradePathInvalid,
-			Message:            fmt.Sprintf("upgrade from %s to %s is not sequential; only sequential upgrades are supported", from.Raw, to.Raw),
-		})
-		r.Recorder.Eventf(keystone, corev1.EventTypeWarning, conditionReasonUpgradePathInvalid, "Upgrade from %s to %s is not sequential", from.Raw, to.Raw)
-		return ctrl.Result{}, fmt.Errorf("upgrade from %s to %s is not sequential; only sequential upgrades are supported", from.Raw, to.Raw)
-	}
-
-	keystone.Status.TargetRelease = keystone.Spec.Image.Tag
-	keystone.Status.UpgradePhase = keystonev1alpha1.UpgradePhaseExpanding
-	conditions.SetCondition(&keystone.Status.Conditions, metav1.Condition{
-		Type:               "DatabaseReady",
-		Status:             metav1.ConditionFalse,
-		ObservedGeneration: keystone.Generation,
-		Reason:             conditionReasonExpandInProgress,
-		Message:            fmt.Sprintf("Upgrade detected: %s \u2192 %s (phase: Expanding)", from.Raw, to.Raw),
-	})
-
-	logger.Info("Upgrade detected", "from", from.Raw, "to", to.Raw, "phase", keystonev1alpha1.UpgradePhaseExpanding)
-	r.Recorder.Eventf(keystone, corev1.EventTypeNormal, conditionReasonUpgradeInitiated, "Upgrade initiated: %s \u2192 %s", from.Raw, to.Raw)
-	return ctrl.Result{Requeue: true}, nil
-}
-
 // abortUpgrade cancels an in-flight upgrade when spec.image.tag is reverted to
 // status.installedRelease. It deletes the expand/migrate/contract Jobs, clears
 // status.upgradePhase/targetRelease, emits an UpgradeAborted event, and requeues
@@ -582,167 +341,6 @@ func (r *KeystoneReconciler) deleteUpgradeJobs(ctx context.Context, keystone *ke
 	return nil
 }
 
-// setUpgradePhaseRunning sets DatabaseReady=False with reason "{phase}InProgress"
-// for an upgrade phase that is currently executing.
-func setUpgradePhaseRunning(keystone *keystonev1alpha1.Keystone, phase string) {
-	conditions.SetCondition(&keystone.Status.Conditions, metav1.Condition{
-		Type:               "DatabaseReady",
-		Status:             metav1.ConditionFalse,
-		ObservedGeneration: keystone.Generation,
-		Reason:             phase + "InProgress",
-		Message:            fmt.Sprintf("%s phase running: %s \u2192 %s", phase, keystone.Status.InstalledRelease, keystone.Status.TargetRelease),
-	})
-}
-
-// setUpgradeJobFailed sets DatabaseReady=False with reason "{phase}Failed"
-// when an upgrade job fails.
-func setUpgradeJobFailed(keystone *keystonev1alpha1.Keystone, phase, jobName string, err error) {
-	conditions.SetCondition(&keystone.Status.Conditions, metav1.Condition{
-		Type:               "DatabaseReady",
-		Status:             metav1.ConditionFalse,
-		ObservedGeneration: keystone.Generation,
-		Reason:             phase + "Failed",
-		Message:            fmt.Sprintf("%s job %s failed: %v", phase, jobName, err),
-	})
-}
-
-// reconcileUpgrade handles phased database migration during an active upgrade.
-// It dispatches to the handler for the current UpgradePhase.
-func (r *KeystoneReconciler) reconcileUpgrade(ctx context.Context, keystone *keystonev1alpha1.Keystone, configMapName string) (ctrl.Result, error) {
-	switch keystone.Status.UpgradePhase {
-	case keystonev1alpha1.UpgradePhaseExpanding:
-		return r.reconcileExpand(ctx, keystone, configMapName)
-	case keystonev1alpha1.UpgradePhaseMigrating:
-		return r.reconcileMigrate(ctx, keystone, configMapName)
-	case keystonev1alpha1.UpgradePhaseRollingUpdate:
-		return r.reconcileRollingUpdate(ctx, keystone)
-	case keystonev1alpha1.UpgradePhaseContracting:
-		return r.reconcileContract(ctx, keystone, configMapName)
-	default:
-		return ctrl.Result{}, fmt.Errorf("unknown upgrade phase %q", keystone.Status.UpgradePhase)
-	}
-}
-
-// reconcileExpand runs the db_sync --expand Job using the NEW image (spec.image.tag).
-//
-// Per the OpenStack Keystone rolling upgrade procedure, expand migrations are
-// executed with the N+1 (target) code: the N+1 alembic tree owns the schema
-// deltas for the upgrade, and running expand with the N (old) binary only
-// advances the expand head to the old release's HEAD. A subsequent contract
-// run with the new binary then fails keystone's _validate_upgrade_order check
-// with "upgrade contract ahead of expand".
-func (r *KeystoneReconciler) reconcileExpand(ctx context.Context, keystone *keystonev1alpha1.Keystone, configMapName string) (ctrl.Result, error) {
-	expandJob := buildExpandJob(keystone, configMapName, keystone.Spec.Image.Tag)
-	done, err := job.RunJob(ctx, r.Client, r.Scheme, keystone, expandJob)
-	// Emit db_sync metrics for the expand-phase Job so the dashboard panel
-	// and failure-rate alerts continue to observe activity during upgrades
-	r.recordDBJobTerminalState(ctx, keystone, "db-expand")
-	if err != nil {
-		setUpgradeJobFailed(keystone, "Expand", expandJob.Name, err)
-		r.Recorder.Eventf(keystone, corev1.EventTypeWarning, conditionReasonExpandFailed, "Expand job %s failed: %v", expandJob.Name, err)
-		return ctrl.Result{}, fmt.Errorf("running expand job: %w", err)
-	}
-	if !done {
-		setUpgradePhaseRunning(keystone, "Expand")
-		return ctrl.Result{RequeueAfter: RequeueUpgradeWait}, nil
-	}
-
-	// Expand complete \u2014 transition to Migrating.
-	keystone.Status.UpgradePhase = keystonev1alpha1.UpgradePhaseMigrating
-	conditions.SetCondition(&keystone.Status.Conditions, metav1.Condition{
-		Type:               "DatabaseReady",
-		Status:             metav1.ConditionFalse,
-		ObservedGeneration: keystone.Generation,
-		Reason:             conditionReasonMigrateInProgress,
-		Message:            fmt.Sprintf("Expand complete, starting migrate: %s \u2192 %s", keystone.Status.InstalledRelease, keystone.Status.TargetRelease),
-	})
-	r.Recorder.Eventf(keystone, corev1.EventTypeNormal, conditionReasonExpandComplete, "Expand phase complete: %s \u2192 %s", keystone.Status.InstalledRelease, keystone.Status.TargetRelease)
-	return ctrl.Result{Requeue: true}, nil
-}
-
-// reconcileMigrate runs the db_sync --migrate Job using the NEW image (spec.image.tag).
-//
-// Like expand, migrate is executed with the N+1 (target) code so that the
-// alembic data-migration steps defined in the new release are applied (see
-// reconcileExpand for the full rationale).
-func (r *KeystoneReconciler) reconcileMigrate(ctx context.Context, keystone *keystonev1alpha1.Keystone, configMapName string) (ctrl.Result, error) {
-	migrateJob := buildMigrateJob(keystone, configMapName, keystone.Spec.Image.Tag)
-	done, err := job.RunJob(ctx, r.Client, r.Scheme, keystone, migrateJob)
-	// Emit db_sync metrics for the migrate-phase Job so the dashboard panel
-	// and failure-rate alerts continue to observe activity during upgrades
-	r.recordDBJobTerminalState(ctx, keystone, "db-migrate")
-	if err != nil {
-		setUpgradeJobFailed(keystone, "Migrate", migrateJob.Name, err)
-		r.Recorder.Eventf(keystone, corev1.EventTypeWarning, conditionReasonMigrateFailed, "Migrate job %s failed: %v", migrateJob.Name, err)
-		return ctrl.Result{}, fmt.Errorf("running migrate job: %w", err)
-	}
-	if !done {
-		setUpgradePhaseRunning(keystone, "Migrate")
-		return ctrl.Result{RequeueAfter: RequeueUpgradeWait}, nil
-	}
-
-	// Migrate complete \u2014 transition to RollingUpdate.
-	keystone.Status.UpgradePhase = keystonev1alpha1.UpgradePhaseRollingUpdate
-	conditions.SetCondition(&keystone.Status.Conditions, metav1.Condition{
-		Type:               "DatabaseReady",
-		Status:             metav1.ConditionFalse,
-		ObservedGeneration: keystone.Generation,
-		Reason:             conditionReasonUpgradeRollingUpdate,
-		Message:            fmt.Sprintf("Migrate complete, waiting for Deployment rollout: %s \u2192 %s", keystone.Status.InstalledRelease, keystone.Status.TargetRelease),
-	})
-	r.Recorder.Eventf(keystone, corev1.EventTypeNormal, conditionReasonMigrateComplete, "Migrate phase complete: %s \u2192 %s", keystone.Status.InstalledRelease, keystone.Status.TargetRelease)
-	return ctrl.Result{Requeue: true}, nil
-}
-
-// reconcileRollingUpdate is a pass-through that sets a condition and returns an empty
-// result so the main reconcile loop proceeds to reconcileDeployment.
-func (r *KeystoneReconciler) reconcileRollingUpdate(ctx context.Context, keystone *keystonev1alpha1.Keystone) (ctrl.Result, error) {
-	conditions.SetCondition(&keystone.Status.Conditions, metav1.Condition{
-		Type:               "DatabaseReady",
-		Status:             metav1.ConditionFalse,
-		ObservedGeneration: keystone.Generation,
-		Reason:             conditionReasonUpgradeRollingUpdate,
-		Message:            fmt.Sprintf("Waiting for Deployment rollout: %s \u2192 %s", keystone.Status.InstalledRelease, keystone.Status.TargetRelease),
-	})
-	return ctrl.Result{}, nil
-}
-
-// reconcileContract runs the db_sync --contract Job using the NEW image (spec.image.tag)
-// and finalizes the upgrade on completion.
-func (r *KeystoneReconciler) reconcileContract(ctx context.Context, keystone *keystonev1alpha1.Keystone, configMapName string) (ctrl.Result, error) {
-	contractJob := buildContractJob(keystone, configMapName, keystone.Spec.Image.Tag)
-	done, err := job.RunJob(ctx, r.Client, r.Scheme, keystone, contractJob)
-	// Emit db_sync metrics for the contract-phase Job so the dashboard panel
-	// and failure-rate alerts continue to observe activity during upgrades
-	r.recordDBJobTerminalState(ctx, keystone, "db-contract")
-	if err != nil {
-		setUpgradeJobFailed(keystone, "Contract", contractJob.Name, err)
-		r.Recorder.Eventf(keystone, corev1.EventTypeWarning, conditionReasonContractFailed, "Contract job %s failed: %v", contractJob.Name, err)
-		return ctrl.Result{}, fmt.Errorf("running contract job: %w", err)
-	}
-	if !done {
-		setUpgradePhaseRunning(keystone, "Contract")
-		return ctrl.Result{RequeueAfter: RequeueUpgradeWait}, nil
-	}
-
-	// Contract complete \u2014 upgrade finished.
-	from := keystone.Status.InstalledRelease
-	to := keystone.Status.TargetRelease
-	keystone.Status.InstalledRelease = to
-	keystone.Status.TargetRelease = ""
-	keystone.Status.UpgradePhase = ""
-	conditions.SetCondition(&keystone.Status.Conditions, metav1.Condition{
-		Type:               "DatabaseReady",
-		Status:             metav1.ConditionTrue,
-		ObservedGeneration: keystone.Generation,
-		Reason:             conditionReasonDatabaseSynced,
-		Message:            fmt.Sprintf("Database schema is up to date (upgraded %s \u2192 %s)", from, to),
-	})
-	r.Recorder.Eventf(keystone, corev1.EventTypeNormal, conditionReasonUpgradeComplete, "Upgrade complete: %s \u2192 %s", from, to)
-	log.FromContext(ctx).Info("Upgrade complete", "from", from, "to", to)
-	return ctrl.Result{}, nil
-}
-
 // finalizeDatabaseResources issues Delete for the MariaDB Database, User, and
 // Grant CRs named after the Keystone CR. It returns as soon as the Delete
 // requests have been accepted (or tolerated as NotFound) so the Keystone CR
@@ -786,194 +384,4 @@ func (r *KeystoneReconciler) finalizeDatabaseResources(ctx context.Context, keys
 	}
 
 	return nil
-}
-
-func buildDatabase(keystone *keystonev1alpha1.Keystone) *mariadbv1alpha1.Database {
-	key := mariaDBResourceKey(keystone)
-	return &mariadbv1alpha1.Database{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      key.Name,
-			Namespace: key.Namespace,
-		},
-		Spec: mariadbv1alpha1.DatabaseSpec{
-			MariaDBRef: mariadbv1alpha1.MariaDBRef{
-				ObjectReference: mariadbv1alpha1.ObjectReference{
-					Name: keystone.Spec.Database.ClusterRef.Name,
-				},
-			},
-			CharacterSet: "utf8",
-			Collate:      "utf8_general_ci",
-			Name:         keystone.Spec.Database.Database,
-		},
-	}
-}
-
-func buildUser(keystone *keystonev1alpha1.Keystone) *mariadbv1alpha1.User {
-	key := mariaDBResourceKey(keystone)
-	return &mariadbv1alpha1.User{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      key.Name,
-			Namespace: key.Namespace,
-		},
-		Spec: mariadbv1alpha1.UserSpec{
-			MariaDBRef: mariadbv1alpha1.MariaDBRef{
-				ObjectReference: mariadbv1alpha1.ObjectReference{
-					Name: keystone.Spec.Database.ClusterRef.Name,
-				},
-			},
-			PasswordSecretKeyRef: &mariadbv1alpha1.SecretKeySelector{
-				LocalObjectReference: mariadbv1alpha1.LocalObjectReference{
-					Name: keystone.Spec.Database.SecretRef.Name,
-				},
-				Key: "password",
-			},
-		},
-	}
-}
-
-func buildGrant(keystone *keystonev1alpha1.Keystone) *mariadbv1alpha1.Grant {
-	key := mariaDBResourceKey(keystone)
-	return &mariadbv1alpha1.Grant{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      key.Name,
-			Namespace: key.Namespace,
-		},
-		Spec: mariadbv1alpha1.GrantSpec{
-			MariaDBRef: mariadbv1alpha1.MariaDBRef{
-				ObjectReference: mariadbv1alpha1.ObjectReference{
-					Name: keystone.Spec.Database.ClusterRef.Name,
-				},
-			},
-			Privileges: []string{"ALL PRIVILEGES"},
-			Database:   keystone.Spec.Database.Database,
-			Table:      "*",
-			Username:   key.Name,
-		},
-	}
-}
-
-// buildDBJob constructs a keystone-manage db_sync Job with the shared container spec,
-// volume mounts, and security context used by both regular db_sync and upgrade phase
-// jobs. This single builder prevents drift when these need to change in the future.
-//
-// TODO Wire spec.Resources (or a smaller Job-specific default) to the
-// container. Currently runs as BestEffort QoS. See reconcile_deployment.go
-// containerResources() for the pattern used by the keystone container.
-func buildDBJob(keystone *keystonev1alpha1.Keystone, configMapName, imageTag, nameSuffix string, command []string) *batchv1.Job {
-	backoffLimit := int32(4)
-	job := &batchv1.Job{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      fmt.Sprintf("%s-%s", keystone.Name, nameSuffix),
-			Namespace: keystone.Namespace,
-		},
-		Spec: batchv1.JobSpec{
-			BackoffLimit: &backoffLimit,
-			Template: corev1.PodTemplateSpec{
-				Spec: corev1.PodSpec{
-					RestartPolicy:     corev1.RestartPolicyNever,
-					PriorityClassName: priorityClassName(keystone),
-					Containers: []corev1.Container{{
-						Name:            nameSuffix,
-						Image:           fmt.Sprintf("%s:%s", keystone.Spec.Image.Repository, imageTag),
-						Command:         command,
-						SecurityContext: restrictedSecurityContext(),
-						// Override [database].connection via oslo.config env-var so every
-						// db_sync variant (db-sync, expand, migrate, contract, schema-check)
-						// reads the DB URL from the derived Secret instead of the ConfigMap
-						Env: []corev1.EnvVar{buildDBConnectionEnvVar(keystone)},
-						VolumeMounts: []corev1.VolumeMount{{
-							Name:      "config",
-							MountPath: "/etc/keystone/keystone.conf.d/",
-							ReadOnly:  true,
-						}},
-					}},
-					Volumes: []corev1.Volume{{
-						Name: "config",
-						VolumeSource: corev1.VolumeSource{
-							ConfigMap: &corev1.ConfigMapVolumeSource{
-								LocalObjectReference: corev1.LocalObjectReference{
-									Name: configMapName,
-								},
-							},
-						},
-					}},
-				},
-			},
-		},
-	}
-	// project the db-tls client keypair into every db_sync variant
-	// (db-sync, expand, migrate, contract, schema-check) when DB TLS is
-	// enabled; the gate is centralised in dbTLSEnabled so deployment and job
-	// builders decide identically.
-	if dbTLSEnabled(keystone) {
-		tlsVol, tlsMount := dbTLSVolumeAndMount(keystone)
-		job.Spec.Template.Spec.Volumes = append(job.Spec.Template.Spec.Volumes, tlsVol)
-		job.Spec.Template.Spec.Containers[0].VolumeMounts = append(
-			job.Spec.Template.Spec.Containers[0].VolumeMounts, tlsMount,
-		)
-	}
-	return job
-}
-
-func buildDBSyncJob(keystone *keystonev1alpha1.Keystone, configMapName string) *batchv1.Job {
-	return buildDBJob(keystone, configMapName, keystone.Spec.Image.Tag, "db-sync",
-		[]string{"keystone-manage", "--config-dir=/etc/keystone/keystone.conf.d/", "db_sync"})
-}
-
-// buildUpgradeJob creates a db_sync Job for one of the expand-migrate-contract
-// upgrade phases. All three phases run with the NEW release image:
-// reconcileExpand, reconcileMigrate, and reconcileContract each pass
-// keystone.Spec.Image.Tag, per the OpenStack rolling-upgrade procedure where
-// the N+1 alembic tree owns the schema deltas (see reconcileExpand for the
-// full rationale). The imageTag parameter is retained so a caller could pin a
-// phase to a different image, but the upgrade flow always passes the target
-// release to all three — do not read this comment as "expand/migrate use the
-// old release" when triaging a stuck upgrade.
-func buildUpgradeJob(keystone *keystonev1alpha1.Keystone, configMapName, imageTag, phase, flag string) *batchv1.Job {
-	return buildDBJob(keystone, configMapName, imageTag, fmt.Sprintf("db-%s", phase),
-		[]string{"keystone-manage", "--config-dir=/etc/keystone/keystone.conf.d/", "db_sync", flag})
-}
-
-// buildExpandJob creates a db_sync --expand Job using the given imageTag.
-func buildExpandJob(keystone *keystonev1alpha1.Keystone, configMapName, imageTag string) *batchv1.Job {
-	return buildUpgradeJob(keystone, configMapName, imageTag, "expand", "--expand")
-}
-
-// buildMigrateJob creates a db_sync --migrate Job using the given imageTag.
-func buildMigrateJob(keystone *keystonev1alpha1.Keystone, configMapName, imageTag string) *batchv1.Job {
-	return buildUpgradeJob(keystone, configMapName, imageTag, "migrate", "--migrate")
-}
-
-// buildContractJob creates a db_sync --contract Job using the given imageTag.
-func buildContractJob(keystone *keystonev1alpha1.Keystone, configMapName, imageTag string) *batchv1.Job {
-	return buildUpgradeJob(keystone, configMapName, imageTag, "contract", "--contract")
-}
-
-// buildSchemaCheckJob constructs a schema-check Job that verifies the database
-// schema matches the expected Alembic migration state after db_sync completes.
-// The Job runs keystone-manage db_sync --check which exits 0 when the schema is
-// up-to-date, and 1..4 when expand/migrate/contract migrations are pending.
-// It delegates to buildDBJob for the shared pod spec and overrides backoffLimit=2
-// for a read-only check. TTLSecondsAfterFinished is
-// intentionally left unset: the completed Job is the RunJob state record, and a
-// TTL-driven garbage-collection would re-create it on the next reconcile, causing
-// a re-creation loop (#415). The Job is cleaned up via owner-reference GC
-// with the Keystone CR.
-func buildSchemaCheckJob(keystone *keystonev1alpha1.Keystone, configMapName string) *batchv1.Job {
-	// Read-only schema verification via keystone-manage db_sync --check.
-	// Exit codes: 0 = up-to-date, 1..4 = needs expand/migrate/contract.
-	// This avoids parsing db_version output, which mixes Oslo log lines with
-	// revision hashes and only reports the expand head (not the contract head).
-	schemaCheckScript := `keystone-manage --config-dir=/etc/keystone/keystone.conf.d/ db_sync --check`
-
-	j := buildDBJob(keystone, configMapName, keystone.Spec.Image.Tag, "schema-check",
-		[]string{"/bin/sh", "-eu", "-c", schemaCheckScript})
-
-	// Override defaults: fewer retries for a read-only check. The completed Job
-	// lingers as the RunJob state record (no TTL) to avoid a TTL-driven
-	// re-creation loop (#415).
-	backoffLimit := int32(2)
-	j.Spec.BackoffLimit = &backoffLimit
-
-	return j
 }
